+++ source/adios2/engine/bp5/BP5Engine.h
@@ -174,7 +174,15 @@ public:
     MACRO(RemoteHost, String, std::string, "")                                                     \
     MACRO(UUID, String, std::string, "")                                                           \
     MACRO(TarInfo, String, std::string, "")                                                        \
-    MACRO(MaxOpenFilesAtOnce, UInt, unsigned int, UINT_MAX)
+    MACRO(MaxOpenFilesAtOnce, UInt, unsigned int, UINT_MAX)                                        \
+    MACRO(DataTransport, String, std::string, "")                                                  \
+    MACRO(S3Endpoint, String, std::string, "")                                                     \
+    MACRO(S3Bucket, String, std::string, "")                                                       \
+    MACRO(S3AccessKeyID, String, std::string, "")                                                  \
+    MACRO(S3SecretKey, String, std::string, "")                                                    \
+    MACRO(S3SessionToken, String, std::string, "")                                                 \
+    MACRO(S3ObjectMode, String, std::string, "multi")                                              \
+    MACRO(S3DirectUploadThreshold, String, std::string, "")
 
     struct BP5Params
     {