+++ source/adios2/engine/bp5/BP5Writer_EveryoneWrites_Async.cpp
@@ -211,6 +211,7 @@ int BP5Writer::AsyncWriteThread_EveryoneWrites(AsyncWriteInfo *info)
             info->tokenChain->RecvToken();
         }
     }
+    info->tm->FinalizeSegment();
     delete info->Data;
     return 1;
 };